+++ Sources/SwifterSwift/Foundation/URLExtensions.swift
@@ -32,6 +32,20 @@ public extension URL {
 
 }
 
+// MARK: - Initializers
+public extension URL {
+
+    /// SwifterSwift: Initializes an `URL` object with a base URL and a relative string. If `string` was malformed, returns `nil`.
+    /// - Parameters:
+    ///   - string: The URL string with which to initialize the `URL` object. Must conform to RFC 2396. `string` is interpreted relative to `url`.
+    ///   - url: The base URL for the `URL` object.
+    init?(string: String?, relativeTo url: URL? = nil) {
+        guard let string = string else { return nil }
+        self.init(string: string, relativeTo: url)
+    }
+
+}
+
 // MARK: - Methods
 public extension URL {